massive refactoring (wip).

+++ geokit/src/units/angle.rs
@@ -1,18 +1,20 @@
-use std::f64::consts::PI;
+use crate::math::{Float, PI};
 
 /// An angle unit as a to-radians angle converter.
 /// An angle expressed in this unit is converted to radians as follow:
 /// ```
+/// use geokit::units::angle::AngleUnit;
+/// use geokit::math::PI;
 /// let deg = AngleUnit(PI, 180.0);
 /// let angle_deg = 1.0;
-/// leet angle_rad = angle_deg * deg.0 / deg.1;
+/// let angle_rad = angle_deg * deg.rad_per_unit();
 /// ```
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub struct AngleUnit(pub f64, pub f64);
+pub struct AngleUnit(pub Float, pub Float);
 
 impl AngleUnit {
     #[inline]
-    pub const fn rad_per_unit(&self) -> f64 {
+    pub const fn rad_per_unit(&self) -> Float {
         self.0 / self.1
     }
 }
